+++ ios/Classes/lib/notifications/models/NotificationContentModel.swift
@@ -21,8 +21,9 @@ public class NotificationContentModel : AbstractModel {
     
     var playSound: Bool?
     var customSound: String?
-    var largeIcon: String?
     var locked: Bool?
+    var icon: String?
+    var largeIcon: String?
     var bigPicture: String?
     var hideLargeIconOnExpand: Bool?
     var autoCancel: Bool?
@@ -55,8 +56,9 @@ public class NotificationContentModel : AbstractModel {
         
         self.playSound             = MapUtils<Bool>.getValueOrDefault(reference: "playSound", arguments: arguments)
         self.customSound           = MapUtils<String>.getValueOrDefault(reference: "customSound", arguments: arguments)
-        self.largeIcon             = MapUtils<String>.getValueOrDefault(reference: "largeIcon", arguments: arguments)
         self.locked                = MapUtils<Bool>.getValueOrDefault(reference: "locked", arguments: arguments)
+        self.icon                  = MapUtils<String>.getValueOrDefault(reference: "icon", arguments: arguments)
+        self.largeIcon             = MapUtils<String>.getValueOrDefault(reference: "largeIcon", arguments: arguments)
         self.bigPicture            = MapUtils<String>.getValueOrDefault(reference: "bigPicture", arguments: arguments)
         self.hideLargeIconOnExpand = MapUtils<Bool>.getValueOrDefault(reference: "hideLargeIconOnExpand", arguments: arguments)
         self.autoCancel            = MapUtils<Bool>.getValueOrDefault(reference: "autoCancel", arguments: arguments)
@@ -120,6 +122,7 @@ public class NotificationContentModel : AbstractModel {
         if(self.showWhen != nil){ mapData["showWhen"] = self.showWhen }
         if(self.playSound != nil){ mapData["playSound"] = self.playSound }
         if(self.customSound != nil){ mapData["customSound"] = self.customSound }
+        if(self.icon != nil){ mapData["icon"] = self.icon }
         if(self.largeIcon != nil){ mapData["largeIcon"] = self.largeIcon }
         if(self.locked != nil){ mapData["locked"] = self.locked }
         if(self.bigPicture != nil){ mapData["bigPicture"] = self.bigPicture }
@@ -154,6 +157,15 @@ public class NotificationContentModel : AbstractModel {
         if(StringUtils.isNullOrEmpty(channelKey)){
             throw PushNotificationError.invalidRequiredFields(
                 msg: "channelKey cannot be null or empty")
+        }
+
+        if(!StringUtils.isNullOrEmpty(icon)){
+            if(
+                BitmapUtils.getMediaSourceType(icon) != MediaSource.Resource
+            ){
+                throw PushNotificationError.invalidRequiredFields(
+                    msg: "Small icon ('"+icon+"') must be a valid media native resource type.")
+            }
         }
         
         if(notificationLayout == nil){